--- proto/Project.toml
@@ -1,19 +1,0 @@
-name = "proto"
-uuid = "66c8cbca-5ec6-4091-aa7b-25ec20c0f1c0"
-authors = ["colinxs <me@colinxsummers.com>"]
-version = "0.1.0"
-
-[deps]
-ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-
-[compat]
-StaticArrays = "^0.12"
-julia = "1.1.1"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
