--- proto/src/proto.jl
@@ -1,5 +1,0 @@
-module proto
-
-greet() = print("Hello World!")
-
-end # module
